requirements.txt
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
werkzeug==2.2.2
psycopg2==2.9.5
python-dotenv==0.21.0
gunicorn==20.1.0
mysqlclient==2.1.0
pylist
pytest
pytest-cov
flask
openai
requests